Game Title: What-The-Fact
This project utilizes the following programming languages: HTML5, CSS, and JavaScript. The title of this game is "What-The-Fact" which enhance the critical thinking and mental memory of a student. The player will choose atleast three (2) categoies that consist of two question. This game are related to STEM that will boost the players mind.